pyproject.toml
[tool.poetry]
name = "spartan-framework"
version = "0.1.8"
description = "A micro framework for serverless development in some cloud providers"
authors = ["Sydel Palinlin <sydel.palinlin@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "spartan_framework"}]

[tool.poetry.dependencies]
python = "^3.11"
alembic = "^1.12.0"
fastapi = "^0.99.1"
uvicorn = "^0.23.2"
httpx = "^0.25.0"
mangum = "^0.17.0"
python-dotenv = "^1.0.0"
boto3 = "^1.28.66"
pg8000 = "^1.30.2"
email-validator = "^2.0.0.post2"
jsonpickle = "^3.0.2"
jinja2 = "^3.1.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
black = "^23.9.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
bandit = "^1.7.5"
poetry-plugin-export = "^1.6.0"
moto = "^4.2.7"
pytest-mock = "^3.12.0"
pytest-asyncio = "^0.21.1"
faker = "^20.1.0"
mocker = "^1.1.1"
ipykernel = "^6.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
